Added support for Bearer token on Webhook class
Webhook class now supports the addition of an optional "token". It needs to be declared at the same level that "url" on the application.conf file. If the parameter exists within a webhook definition, its value will be added to an Authorization header. Else, the send() function of the Webhook class won't include this header on the POST request.

An example of configuration on the application.conf would be as follows:
webhooks {
  myLocalWebHook {
    url = "http://localhost/webhook"
    token = "xxxxxxxxxxxxxxxxxx"
  }
}